Refactor Inventory class from inventory directory to model directory
