+++ src/main/java/nustorage/model/ModelManager.java
@@ -14,7 +14,6 @@ import javafx.collections.transformation.FilteredList;
 import nustorage.commons.core.GuiSettings;
 import nustorage.commons.core.LogsCenter;
 import nustorage.commons.core.index.Index;
-import nustorage.model.inventory.Inventory;
 import nustorage.model.person.Person;
 import nustorage.model.record.FinanceRecord;
 import nustorage.model.record.InventoryRecord;
