+++ src/main/java/nustorage/model/ReadOnlyInventory.java
@@ -1,4 +1,4 @@
-package nustorage.model.inventory;
+package nustorage.model;
 
 import javafx.collections.ObservableList;
 import nustorage.model.record.InventoryRecord;
