+++ src/main/java/nustorage/model/record/exceptions/NegativeNumberOfItemException.java
@@ -1,4 +1,4 @@
-package nustorage.model.inventory.exceptions;
+package nustorage.model.record.exceptions;
 
 import nustorage.model.record.InventoryRecord;
